fix: commit suggestions pt.3 remove appstore icons and move functions to constant file
